Window frame state capturing the state of the window frame calculations
