+++ datafusion/physical-expr/src/window/aggregate.rs
@@ -31,7 +31,9 @@ use datafusion_common::ScalarValue;
 use datafusion_expr::WindowFrame;
 
 use crate::{expressions::PhysicalSortExpr, PhysicalExpr};
-use crate::{window::groups::WindowFrameGroups, window::WindowExpr, AggregateExpr};
+use crate::{window::WindowExpr, AggregateExpr};
+
+use super::window_frame_state::{WindowFrameState, WindowFrameStateImpl};
 
 /// A window expr that takes the form of an aggregate function
 #[derive(Debug)]
@@ -114,15 +116,14 @@ impl WindowExpr for AggregateWindowExpr {
                 .map(|v| v.slice(partition_range.start, length))
                 .collect::<Vec<_>>();
 
-            let mut window_frame_groups = WindowFrameGroups::default();
+            let mut window_frame_state = WindowFrameState::new(&window_frame);
             let mut last_range: (usize, usize) = (0, 0);
 
             // We iterate on each row to perform a running calculation.
             // First, cur_range is calculated, then it is compared with last_range.
             for i in 0..length {
-                let cur_range = self.calculate_range(
+                let cur_range = window_frame_state.calculate_range(
                     &window_frame,
-                    &mut window_frame_groups,
                     &slice_order_bys,
                     &sort_options,
                     length,
