+++ datafusion/physical-expr/src/window/mod.rs
@@ -19,13 +19,13 @@ mod aggregate;
 mod built_in;
 mod built_in_window_function_expr;
 pub(crate) mod cume_dist;
-mod groups;
 pub(crate) mod lead_lag;
 pub(crate) mod nth_value;
 pub(crate) mod partition_evaluator;
 pub(crate) mod rank;
 pub(crate) mod row_number;
 mod window_expr;
+mod window_frame_state;
 
 pub use aggregate::AggregateWindowExpr;
 pub use built_in::BuiltInWindowExpr;
