+++ datafusion/physical-expr/src/window/window_expr.rs
@@ -20,18 +20,11 @@ use arrow::compute::kernels::partition::lexicographical_partition_ranges;
 use arrow::compute::kernels::sort::{SortColumn, SortOptions};
 use arrow::record_batch::RecordBatch;
 use arrow::{array::ArrayRef, datatypes::Field};
-use datafusion_common::bisect::bisect;
-use datafusion_common::{DataFusionError, Result, ScalarValue};
+use datafusion_common::{DataFusionError, Result};
 use std::any::Any;
-use std::cmp::min;
 use std::fmt::Debug;
 use std::ops::Range;
 use std::sync::Arc;
-
-use datafusion_expr::WindowFrameBound;
-use datafusion_expr::{WindowFrame, WindowFrameUnits};
-
-use crate::window::groups::WindowFrameGroups;
 
 /// A window expression that:
 /// * knows its resulting field
@@ -117,317 +110,4 @@ pub trait WindowExpr: Send + Sync + Debug {
         sort_columns.extend(order_by_columns);
         Ok(sort_columns)
     }
-
-    /// We use start and end bounds to calculate current row's starting and ending range.
-    /// This function supports different modes.
-    fn calculate_range(
-        &self,
-        window_frame: &Option<Arc<WindowFrame>>,
-        window_frame_groups: &mut WindowFrameGroups,
-        range_columns: &[ArrayRef],
-        sort_options: &[SortOptions],
-        length: usize,
-        idx: usize,
-    ) -> Result<(usize, usize)> {
-        if let Some(window_frame) = window_frame {
-            match window_frame.units {
-                WindowFrameUnits::Range => {
-                    let start = match &window_frame.start_bound {
-                        WindowFrameBound::Preceding(n) => {
-                            if n.is_null() {
-                                // UNBOUNDED PRECEDING
-                                0
-                            } else {
-                                calculate_index_of_row::<true, true>(
-                                    range_columns,
-                                    sort_options,
-                                    idx,
-                                    Some(n),
-                                )?
-                            }
-                        }
-                        WindowFrameBound::CurrentRow => {
-                            if range_columns.is_empty() {
-                                0
-                            } else {
-                                calculate_index_of_row::<true, true>(
-                                    range_columns,
-                                    sort_options,
-                                    idx,
-                                    None,
-                                )?
-                            }
-                        }
-                        WindowFrameBound::Following(n) => {
-                            calculate_index_of_row::<true, false>(
-                                range_columns,
-                                sort_options,
-                                idx,
-                                Some(n),
-                            )?
-                        }
-                    };
-                    let end = match &window_frame.end_bound {
-                        WindowFrameBound::Preceding(n) => {
-                            calculate_index_of_row::<false, true>(
-                                range_columns,
-                                sort_options,
-                                idx,
-                                Some(n),
-                            )?
-                        }
-                        WindowFrameBound::CurrentRow => {
-                            if range_columns.is_empty() {
-                                length
-                            } else {
-                                calculate_index_of_row::<false, false>(
-                                    range_columns,
-                                    sort_options,
-                                    idx,
-                                    None,
-                                )?
-                            }
-                        }
-                        WindowFrameBound::Following(n) => {
-                            if n.is_null() {
-                                // UNBOUNDED FOLLOWING
-                                length
-                            } else {
-                                calculate_index_of_row::<false, false>(
-                                    range_columns,
-                                    sort_options,
-                                    idx,
-                                    Some(n),
-                                )?
-                            }
-                        }
-                    };
-                    Ok((start, end))
-                }
-                WindowFrameUnits::Rows => {
-                    let start = match window_frame.start_bound {
-                        // UNBOUNDED PRECEDING
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(None)) => 0,
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(Some(n))) => {
-                            if idx >= n as usize {
-                                idx - n as usize
-                            } else {
-                                0
-                            }
-                        }
-                        WindowFrameBound::CurrentRow => idx,
-                        // UNBOUNDED FOLLOWING
-                        WindowFrameBound::Following(ScalarValue::UInt64(None)) => {
-                            return Err(DataFusionError::Internal(format!(
-                                "Frame start cannot be UNBOUNDED FOLLOWING '{:?}'",
-                                window_frame
-                            )))
-                        }
-                        WindowFrameBound::Following(ScalarValue::UInt64(Some(n))) => {
-                            min(idx + n as usize, length)
-                        }
-                        // ERRONEOUS FRAMES
-                        WindowFrameBound::Preceding(_)
-                        | WindowFrameBound::Following(_) => {
-                            return Err(DataFusionError::Internal(
-                                "Rows should be Uint".to_string(),
-                            ))
-                        }
-                    };
-                    let end = match window_frame.end_bound {
-                        // UNBOUNDED PRECEDING
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(None)) => {
-                            return Err(DataFusionError::Internal(format!(
-                                "Frame end cannot be UNBOUNDED PRECEDING '{:?}'",
-                                window_frame
-                            )))
-                        }
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(Some(n))) => {
-                            if idx >= n as usize {
-                                idx - n as usize + 1
-                            } else {
-                                0
-                            }
-                        }
-                        WindowFrameBound::CurrentRow => idx + 1,
-                        // UNBOUNDED FOLLOWING
-                        WindowFrameBound::Following(ScalarValue::UInt64(None)) => length,
-                        WindowFrameBound::Following(ScalarValue::UInt64(Some(n))) => {
-                            min(idx + n as usize + 1, length)
-                        }
-                        // ERRONEOUS FRAMES
-                        WindowFrameBound::Preceding(_)
-                        | WindowFrameBound::Following(_) => {
-                            return Err(DataFusionError::Internal(
-                                "Rows should be Uint".to_string(),
-                            ))
-                        }
-                    };
-                    Ok((start, end))
-                }
-                WindowFrameUnits::Groups => {
-                    if range_columns.is_empty() {
-                        return Err(DataFusionError::Execution(
-                            "GROUPS mode requires an ORDER BY clause".to_string(),
-                        ));
-                    }
-                    let start = match window_frame.start_bound {
-                        // UNBOUNDED PRECEDING
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(None)) => 0,
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(Some(n))) => {
-                            calculate_index_of_group::<true, true>(
-                                range_columns,
-                                window_frame_groups,
-                                idx,
-                                n,
-                                length,
-                            )?
-                        }
-                        WindowFrameBound::CurrentRow => {
-                            calculate_index_of_group::<true, true>(
-                                range_columns,
-                                window_frame_groups,
-                                idx,
-                                0,
-                                length,
-                            )?
-                        }
-                        WindowFrameBound::Following(ScalarValue::UInt64(Some(n))) => {
-                            calculate_index_of_group::<true, false>(
-                                range_columns,
-                                window_frame_groups,
-                                idx,
-                                n,
-                                length,
-                            )?
-                        }
-                        // UNBOUNDED FOLLOWING
-                        WindowFrameBound::Following(ScalarValue::UInt64(None)) => {
-                            return Err(DataFusionError::Internal(format!(
-                                "Frame start cannot be UNBOUNDED FOLLOWING '{:?}'",
-                                window_frame
-                            )))
-                        }
-                        // ERRONEOUS FRAMES
-                        WindowFrameBound::Preceding(_)
-                        | WindowFrameBound::Following(_) => {
-                            return Err(DataFusionError::Internal(
-                                "Groups should be Uint".to_string(),
-                            ))
-                        }
-                    };
-                    let end = match window_frame.end_bound {
-                        // UNBOUNDED PRECEDING
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(None)) => {
-                            return Err(DataFusionError::Internal(format!(
-                                "Frame end cannot be UNBOUNDED PRECEDING '{:?}'",
-                                window_frame
-                            )))
-                        }
-                        WindowFrameBound::Preceding(ScalarValue::UInt64(Some(n))) => {
-                            calculate_index_of_group::<false, true>(
-                                range_columns,
-                                window_frame_groups,
-                                idx,
-                                n,
-                                length,
-                            )?
-                        }
-                        WindowFrameBound::CurrentRow => {
-                            calculate_index_of_group::<false, false>(
-                                range_columns,
-                                window_frame_groups,
-                                idx,
-                                0,
-                                length,
-                            )?
-                        }
-                        WindowFrameBound::Following(ScalarValue::UInt64(Some(n))) => {
-                            calculate_index_of_group::<false, false>(
-                                range_columns,
-                                window_frame_groups,
-                                idx,
-                                n,
-                                length,
-                            )?
-                        }
-                        // UNBOUNDED FOLLOWING
-                        WindowFrameBound::Following(ScalarValue::UInt64(None)) => length,
-                        // ERRONEOUS FRAMES
-                        WindowFrameBound::Preceding(_)
-                        | WindowFrameBound::Following(_) => {
-                            return Err(DataFusionError::Internal(
-                                "Groups should be Uint".to_string(),
-                            ))
-                        }
-                    };
-                    Ok((start, end))
-                }
-            }
-        } else {
-            Ok((0, length))
-        }
-    }
-}
-
-fn calculate_index_of_row<const BISECT_SIDE: bool, const SEARCH_SIDE: bool>(
-    range_columns: &[ArrayRef],
-    sort_options: &[SortOptions],
-    idx: usize,
-    delta: Option<&ScalarValue>,
-) -> Result<usize> {
-    let current_row_values = range_columns
-        .iter()
-        .map(|col| ScalarValue::try_from_array(col, idx))
-        .collect::<Result<Vec<ScalarValue>>>()?;
-    let end_range = if let Some(delta) = delta {
-        let is_descending: bool = sort_options
-            .first()
-            .ok_or_else(|| DataFusionError::Internal("Array is empty".to_string()))?
-            .descending;
-
-        current_row_values
-            .iter()
-            .map(|value| {
-                if value.is_null() {
-                    return Ok(value.clone());
-                }
-                if SEARCH_SIDE == is_descending {
-                    // TODO: Handle positive overflows
-                    value.add(delta)
-                } else if value.is_unsigned() && value < delta {
-                    // NOTE: This gets a polymorphic zero without having long coercion code for ScalarValue.
-                    //       If we decide to implement a "default" construction mechanism for ScalarValue,
-                    //       change the following statement to use that.
-                    value.sub(value)
-                } else {
-                    // TODO: Handle negative overflows
-                    value.sub(delta)
-                }
-            })
-            .collect::<Result<Vec<ScalarValue>>>()?
-    } else {
-        current_row_values
-    };
-    // `BISECT_SIDE` true means bisect_left, false means bisect_right
-    bisect::<BISECT_SIDE>(range_columns, &end_range, sort_options)
-}
-
-fn calculate_index_of_group<const BISECT_SIDE: bool, const SEARCH_SIDE: bool>(
-    range_columns: &[ArrayRef],
-    window_frame_groups: &mut WindowFrameGroups,
-    idx: usize,
-    delta: u64,
-    length: usize,
-) -> Result<usize> {
-    let current_row_values = range_columns
-        .iter()
-        .map(|col| ScalarValue::try_from_array(col, idx))
-        .collect::<Result<Vec<ScalarValue>>>()?;
-    window_frame_groups.process::<BISECT_SIDE, SEARCH_SIDE>(
-        &current_row_values,
-        delta,
-        range_columns,
-        length,
-    )
 }
